#1467 - Modularize agreement
- Move inner classes to own files
- Clean up code in argreement page a bit
- Remove unused code
